requirements.txt
certifi==2024.8.30
cloudinary==1.41.0
python-decouple==3.8
six==1.16.0
urllib3==2.2.3
